# docs/specifications/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Dog Walking Management API
  description: |
    The Stardogwalker management API — Cardiff, South Wales.

    This API is designed **API-first** to support multiple interaction
    channels: web app, native mobile app, conversational chat interfaces
    (e.g. WhatsApp bot), and agentic AI applications. All business logic
    lives here; no channel-specific logic is embedded in the API.

    **Design principles:**
    - Stateless — every request authenticates via JWT bearer token
    - Fine-grained actions — discrete endpoints for each state transition
      (e.g. `/accept`, `/decline`) to support unambiguous intent from any
      client type including AI agents
    - Consistent response envelopes — list responses always include
      `data[]` and `pagination`; errors always return `code`, `message`,
      and optional `details`
    - Versioned — breaking changes will be introduced under `/v2` without
      removing `/v1` until all clients have migrated

    Covers: interest registration (public), authentication, dog owner and
    walker management, dog profiles, walk requests, walk scheduling and
    execution, real-time walk updates, and invoicing.
  version: 1.0.0
  contact:
    name: Dog Walking API Support
    email: support@dogwalking.example.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000/v1
    description: Local development server
  - url: https://api.dogwalking.example.com/v1
    description: Production server
...
security:
  - bearerAuth: []
...
paths:
...
  /walks:
    get:
      tags:
        - Walks
      summary: List walks
      description: |
        Retrieve a paginated list of scheduled or completed walks. Dog owners
        see walks for their dogs; walkers see their assigned walks.
      operationId: listWalks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          description: Filter walks by status
          schema:
            $ref: '#/components/schemas/WalkStatus'
        - name: date
          in: query
          description: Filter walks by date (YYYY-MM-DD)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: A list of walks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkList'
        '401':
          $ref: '#/components/responses/Unauthorised'
  /walks/{walkId}:
    parameters:
      - $ref: '#/components/parameters/WalkId'
    get:
      tags:
        - Walks
      summary: Get a walk
      description: Retrieve details of a specific walk.
      operationId: getWalk
      responses:
        '200':
          description: Walk details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walk'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Walks
      summary: Update a walk
      description: Update details of a scheduled walk (e.g. start time, route notes).
      operationId: updateWalk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkEdit'
      responses:
        '200':
          description: Walk updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /walks/{walkId}/start:
    parameters:
      - $ref: '#/components/parameters/WalkId'
    post:
      tags:
        - Walks
      summary: Start a walk
      description: Mark a scheduled walk as started, transitioning it to `in_progress`.
      operationId: startWalk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkStart'
      responses:
        '200':
          description: Walk started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /walks/{walkId}/complete:
    parameters:
      - $ref: '#/components/parameters/WalkId'
    post:
      tags:
        - Walks
      summary: Complete a walk
      description: Mark a walk as completed.
      operationId: completeWalk
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkComplete'
      responses:
        '200':
          description: Walk marked as completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /walks/{walkId}/cancel:
    parameters:
      - $ref: '#/components/parameters/WalkId'
    post:
      tags:
        - Walks
      summary: Cancel a walk
      description: Cancel a scheduled walk.
      operationId: cancelWalk
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalkCancel'
      responses:
        '200':
          description: Walk cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Walk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Walk Updates (images and notes)
  # ---------------------------------------------------------------------------
  /walks/{walkId}/updates:
    parameters:
      - $ref: '#/components/parameters/WalkId'
    get:
      tags:
        - Walk Updates
      summary: List walk updates
      description: Retrieve all updates (notes and images) posted during a walk.
      operationId: listWalkUpdates
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: A list of walk updates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkUpdateList'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Walk Updates
      summary: Add a walk update
      description: |
        Post a new update for a walk. Supports text notes and image uploads
        via multipart form data.
      operationId: createWalkUpdate
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/WalkUpdateCreate'
      responses:
        '201':
          description: Walk update created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkUpdateItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /walks/{walkId}/updates/{updateId}:
    parameters:
      - $ref: '#/components/parameters/WalkId'
      - $ref: '#/components/parameters/UpdateId'
    get:
      tags:
        - Walk Updates
      summary: Get a walk update
      description: Retrieve a specific walk update.
      operationId: getWalkUpdate
      responses:
        '200':
          description: Walk update details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalkUpdateItem'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Walk Updates
      summary: Delete a walk update
      description: Remove a walk update.
      operationId: deleteWalkUpdate
      responses:
        '204':
          description: Walk update deleted
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Recurring Walks
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    WalkId:
      name: walkId
      in: path
      required: true
      description: Unique identifier of the walk
      schema:
        type: string
        format: uuid
    UpdateId:
      name: updateId
      in: path
      required: true
      description: Unique identifier of the walk update
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Number of results per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validationError:
              summary: Validation error
              value:
                code: VALIDATION_ERROR
                message: One or more fields failed validation.
                details:
                  - field: email
                    issue: Must be a valid email address.
                  - field: durationMinutes
                    issue: Must be between 15 and 240.
            noMatchingRate:
              summary: No matching rate found
              value:
                code: NO_MATCHING_RATE
                message: No rate found for this walk type, dog count, and duration. Please update the walker's rate card.
            invalidStateTransition:
              summary: Invalid state transition
              value:
                code: INVALID_STATE_TRANSITION
                message: This action cannot be performed on a resource in its current status.
            schedulingConflict:
              summary: Scheduling conflict
              value:
                code: SCHEDULING_CONFLICT
                message: The walker already has an overlapping walk scheduled for this time.
    Unauthorised:
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingToken:
              summary: Missing bearer token
              value:
                code: AUTHENTICATION_REQUIRED
                message: A valid bearer token is required to access this resource.
            invalidCredentials:
              summary: Invalid login credentials
              value:
                code: INVALID_CREDENTIALS
                message: The email or password provided is incorrect.
            tokenExpired:
              summary: Expired token
              value:
                code: TOKEN_EXPIRED
                message: The access token has expired. Please refresh your session.
    Forbidden:
      description: Authenticated user does not have permission to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            insufficientRole:
              summary: Insufficient role
              value:
                code: FORBIDDEN
                message: Your role does not have permission to perform this action.
            notOwnResource:
              summary: Not own resource
              value:
                code: FORBIDDEN
                message: You can only access your own records.
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            resourceNotFound:
              summary: Resource not found
              value:
                code: RESOURCE_NOT_FOUND
                message: The requested resource could not be found.
...
  schemas:
...
    Error:
      type: object
      description: |
        Standard error envelope returned by all endpoints on failure.

        **Known error codes:**
        - `VALIDATION_ERROR` – One or more request fields failed validation (see `details`).
        - `NO_MATCHING_RATE` – No walker rate card entry matches the requested walk type, dog count, and duration.
        - `SCHEDULING_CONFLICT` – The walker has an overlapping walk already scheduled.
        - `INVALID_STATE_TRANSITION` – The action is not allowed given the resource's current status.
        - `AUTHENTICATION_REQUIRED` – No valid bearer token was provided.
        - `INVALID_CREDENTIALS` – Email or password is incorrect.
        - `TOKEN_EXPIRED` – The access token has expired.
        - `FORBIDDEN` – The authenticated user does not have permission.
        - `RESOURCE_NOT_FOUND` – The requested resource does not exist.
        - `DUPLICATE_EMAIL` – A user with this email already exists.
        - `ALREADY_PROCESSED` – The resource has already been accepted or declined.
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code (see list above)
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error description
          example: One or more fields failed validation.
        details:
          type: array
          description: Optional field-level error details, typically present for validation errors
          items:
            type: object
            properties:
              field:
                type: string
                description: The request field that failed validation
                example: email
              issue:
                type: string
                description: What is wrong with the field
                example: Must be a valid email address.
    Pagination:
      type: object
      required:
        - page
        - pageSize
        - total
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          description: Total number of matching records
          example: 42
...
    WalkType:
      type: string
      enum:
        - solo_walk
        - group_walk
        - adventure_walk
        - puppy_walk
      description: |
        The format of the walk. Together with the number of dogs and duration,
        this determines which rate from the walker's rate card applies.
      example: solo_walk
...
    WalkStatus:
      type: string
      enum:
        - scheduled
        - in_progress
        - completed
        - cancelled
      example: scheduled
    WalkEdit:
      type: object
      properties:
        scheduledDate:
          type: string
          format: date
          example: '2025-04-10'
        scheduledStartTime:
          type: string
          format: time
          example: '09:15:00Z'
        durationMinutes:
          type: integer
          minimum: 15
          maximum: 240
          example: 60
        routeNotes:
          type: string
          description: Notes about the planned walking route
          example: Use the river path today.
    WalkStart:
      type: object
      required:
        - actualStartTime
      properties:
        actualStartTime:
          type: string
          format: date-time
          description: The actual time the walk started
          example: '2025-04-10T09:01:00Z'
    WalkComplete:
      type: object
      properties:
        actualStartTime:
          type: string
          format: date-time
          example: '2025-04-10T09:02:00Z'
        actualEndTime:
          type: string
          format: date-time
          example: '2025-04-10T10:05:00Z'
        distanceKm:
          type: number
          format: float
          description: Distance walked in kilometres
          example: 3.2
        summaryNotes:
          type: string
          description: Walker's summary of the walk
          example: All dogs were very well behaved today!
    WalkCancel:
      type: object
      properties:
        reason:
          type: string
          description: Reason for cancelling the walk
          example: Dog owner had an emergency.
    Walk:
      type: object
      required:
        - id
        - requestId
        - ownerId
        - walkerId
        - dogIds
        - scheduledDate
        - scheduledStartTime
        - durationMinutes
        - walkType
        - agreedRate
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: dd00ee11-ff22-aa33-bb44-cc5566778899
        requestId:
          type: string
          format: uuid
          example: aa11bb22-cc33-dd44-ee55-ff6677889900
        ownerId:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        walkerId:
          type: string
          format: uuid
          example: 7b1e4d2a-1234-5678-abcd-ef0123456789
        dogIds:
          type: array
          items:
            type: string
            format: uuid
          example:
            - c1a2b3d4-5678-90ab-cdef-1234567890ab
        scheduledDate:
          type: string
          format: date
          example: '2025-04-10'
        scheduledStartTime:
          type: string
          format: time
          example: '09:00:00Z'
        durationMinutes:
          type: integer
          example: 60
        walkType:
          $ref: '#/components/schemas/WalkType'
        agreedRate:
          type: number
          format: float
          description: |
            Hourly rate in GBP (matching `WalkRate.ratePerHour`) resolved from
            the walker's rate card at the time the walk request was accepted,
            based on walkType, number of dogs, and durationMinutes. The total
            walk cost is derived at invoicing time (quantity × unitPrice).
          example: 18.00
        status:
          $ref: '#/components/schemas/WalkStatus'
        routeNotes:
          type: string
          nullable: true
          example: Use the river path today.
        actualStartTime:
          type: string
          format: date-time
          nullable: true
          example: null
        actualEndTime:
          type: string
          format: date-time
          nullable: true
          example: null
        distanceKm:
          type: number
          format: float
          nullable: true
          example: null
        summaryNotes:
          type: string
          nullable: true
          example: null
        cancelReason:
          type: string
          nullable: true
          example: null
        createdAt:
          type: string
          format: date-time
          example: '2025-04-01T15:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2025-04-01T15:00:00Z'
    WalkList:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Walk'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # --- Walk Updates ---------------------------------------------------------
    WalkUpdateCreate:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - note
            - image
          description: Type of update
          example: note
        note:
          type: string
          description: Text note about the walk (required when type is 'note')
          example: Max is having a great time chasing squirrels!
        image:
          type: string
          format: binary
          description: Image file upload (required when type is 'image')
        imageCaption:
          type: string
          description: Optional caption for the uploaded image
          example: Max splashing in the river!
    WalkUpdateItem:
      type: object
      required:
        - id
        - walkId
        - type
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: 11223344-5566-7788-99aa-bbccddeeff00
        walkId:
          type: string
          format: uuid
          example: dd00ee11-ff22-aa33-bb44-cc5566778899
        type:
          type: string
          enum:
            - note
            - image
          example: note
        note:
          type: string
          nullable: true
          example: Max is having a great time chasing squirrels!
        imageUrl:
          type: string
          format: uri
          nullable: true
          description: URL of the uploaded image
          example: https://storage.dogwalking.example.com/walks/dd00ee11/max-river.jpg
        imageCaption:
          type: string
          nullable: true
          example: Max splashing in the river!
        createdAt:
          type: string
          format: date-time
          example: '2025-04-10T09:30:00Z'
    WalkUpdateList:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/WalkUpdateItem'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # --- Invoices -------------------------------------------------------------
...